# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/edge/browse": {
            "get": {
                "description": "取得所有Browse name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "設備列表"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "設備ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/devicerouter.errMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/devicerouter.errMsg"
                        }
                    }
                }
            }
        },
        "/edge/checkconverter": {
            "post": {
                "description": "檢查轉換表是否被編輯",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "設備列表"
                ],
                "parameters": [
                    {
                        "description": "轉換表資料",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/devicerouter.edConvInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/devicerouter.errMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/devicerouter.errMsg"
                        }
                    }
                }
            }
        },
...
        "/edge/converter": {
            "get": {
                "description": "取得轉換表資訊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "設備列表"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "設備ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/nodeattr.Converter"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/devicerouter.errMsg"
                        }
                    }
                }
            },
            "put": {
                "description": "編輯資料轉換表資訊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "設備列表"
                ],
                "parameters": [
                    {
                        "description": "資料轉換表資料",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/devicerouter.edConvInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/devicerouter.errMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/devicerouter.errMsg"
                        }
                    }
                }
            },
            "post": {
                "description": "複製轉換表為新範本(data需提供轉換表所有欄位資料，name填入新範本名稱)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "設備列表"
                ],
                "parameters": [
                    {
                        "description": "轉換表所有欄位資料",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/devicerouter.edConvInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/devicerouter.errMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/devicerouter.errMsg"
                        }
                    }
                }
            }
        },
        "/edge/device": {
            "get": {
                "description": "取得設備資訊\u003cbr\u003e Protocol(value=0:OPC UA,1:HTTP) \u003cbr\u003e若protocol=0 status(value=0:異常,1:運轉,2:閒置,3:Unknown) \u003cbr\u003e若protocol=1 status(value=2:停機,3:閒置,4:異常,5:運轉,7:不良品,其他數字:Unknown)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "設備列表"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/nodeattr.DevInfo"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/devicerouter.errMsg"
                        }
                    }
                }
            },
            "put": {
                "description": "編輯設備資訊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "設備列表"
                ],
                "parameters": [
                    {
                        "description": "設備編輯資料",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/devicerouter.devInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/devicerouter.errMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/devicerouter.errMsg"
                        }
                    }
                }
            },
            "delete": {
                "description": "刪除設備資訊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "設備列表"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "設備ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/devicerouter.errMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/devicerouter.errMsg"
                        }
                    }
                }
            }
        },
...
        "/edge/value": {
            "get": {
                "description": "取得欄位實際值",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "設備列表"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "欄位NodeId",
                        "name": "nodeId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "欄位namespace",
                        "name": "namespace",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "欄位browseName",
                        "name": "browseName",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "設備ID",
                        "name": "devId",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/devicerouter.errMsg"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "devicerouter.devInfo": {
            "type": "object",
            "properties": {
                "editContent": {
                    "$ref": "#/definitions/nodeattr.DevInfo"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "devicerouter.edConvInfo": {
            "type": "object",
            "properties": {
                "editContent": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/nodeattr.Converter"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "devicerouter.errMsg": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "nodeattr.Converter": {
            "type": "object",
            "properties": {
                "convertFunc": {
                    "type": "string"
                },
                "dstBrowseName": {
                    "type": "string"
                },
                "dstNodeId": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "modify": {
                    "type": "integer"
                },
                "referenceBrowseName1": {
                    "type": "string"
                },
                "referenceBrowseName2": {
                    "type": "string"
                },
                "referenceBrowseName3": {
                    "type": "string"
                },
                "srcBrowseName": {
                    "type": "string"
                },
                "srcDevName": {
                    "type": "string"
                },
                "srcNamespace": {
                    "type": "integer"
                },
                "srcNodeId": {
                    "type": "string"
                },
                "srcUnit": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "nodeattr.DevInfo": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "conId": {
                    "type": "integer"
                },
                "conName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "protocol": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "tempID": {
                    "type": "integer"
                },
                "tempName": {
                    "type": "string"
                }
            }
        },
...
    }
}